Añadir clase DistribuidoUtils.php con su correspondiente contenido
